feat: update reminder routes and refactor date functions for improved handling
