+++ src/stores/ReminderStore.ts
@@ -44,7 +44,7 @@ export const useReminderStore = create(
           ...current,
           ...payload,
           title: payload.title !== undefined ? payload.title.trim() : current.title,
-          enabled: true,
+          enabled: payload.enabled ?? current.enabled,
           updatedAt: nowISO(),
         }
